chore(sessions): consolidate boot recovery files
